% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splines.R
\name{deboor}
\alias{deboor}
\title{De Boor's algorithm for evaluating B-splines}
\usage{
deboor(x, t, y, order)
}
\arguments{
\item{x}{A n by 2 matrix of control points.}

\item{t}{A vector of support points}

\item{y}{A vector of evaluation points. Should be sorted!}

\item{order}{The spline order. Cubic splines is order 4.}
}
\value{
A 2 by \code{length(y)} matrix of the evaluated points on the spline.
}
\description{
Evaluate piece wise B-splines a various parameter values. Internal function.
}
\examples{
order <- 3
x <- matrix(rnorm(12), 6, 2)
t <- seq(0, 1, l = nrow(x) - order + 2)
y <- seq(0, 1, l = 50)
print(res <- HierarchicalEdgeBundles:::deboor(x, t, y, order))
plot(x, type = "b", col = "grey", pch = 16)
points(res, col = "red", type = "l", pch = 16, lwd = 3)
}
\keyword{internal}

