% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotHierarchicalEdgeBundles}
\alias{plotHierarchicalEdgeBundles}
\title{Hierarchical Edge Bundling}
\usage{
plotHierarchicalEdgeBundles(phylo, graph, beta = 0.8, include.mrca = FALSE,
  simplify = FALSE, ..., args.lines = list(), args.points = list(pch = 16,
  cex = 0.1), debug = FALSE, e.cols, v.use.only, e.use.only)
}
\arguments{
\item{phylo}{A \code{phylo} object defining the hierarchy.}

\item{graph}{A \code{igraph} object to be drawn.}

\item{beta}{A number between 0 and 1 controlling the bundling strength.}

\item{include.mrca}{Should the only the most recent common ancestor be used
in the shortest path used for the splines?}

\item{simplify}{Simplify the paths by taking the convex hull of the control
points. Can sometimes yield better results.}

\item{...}{Arguments passed to \code{\link[ape]{plot.phylo}}.}

\item{args.lines}{A list of arguments passed to \code{lines}.}

\item{args.points}{A list of arguments passed to \code{points}.}

\item{debug}{Plot some extra debug info.}

\item{e.cols}{A character vector giving the colors of the edges.
Overrides any use of \code{args.lines$col}.}

\item{v.use.only}{An integer vector giving the nodes from which edges are
to be drawn. E.g. \code{use.only = 1} will only plot the edges from
vertex 1.}

\item{e.use.only}{An integer vector giving the edges to be drawn.
E.g. \code{use.only = 1} will only draw the first edge.}
}
\value{
Plots to a new device.
}
\description{
Visualization of networks using hierarchical edge bundles referenced below.
Plots a graph using a hierachical tree as guides for the edges.
}
\examples{
library("igraph")
library("ape")
n <- 10

# Create a graph
corr <- cor(matrix(rnorm(n^2), n, n))
rownames(corr) <- colnames(corr) <-
  apply(combn(LETTERS, 2), 2, paste0, collapse = "")[1:n]
adj <- abs(corr)
graph <- graph.adjacency(adj, mode = "un", weighted = TRUE, diag = FALSE)
E(graph)$color <- ifelse(get.lower.tri(corr) < 0, "blue", "red")

phylo <- as.phylo(hclust(as.dist(1 - adj), method = "ward.D"))
plotHierarchicalEdgeBundles(phylo, graph, type = "fan",
                            e.cols = E(graph)$color)

par(mfrow = c(1, 3), mar = c(0, 0, 0, 0))
plot(phylo, type = "fan")
plotHierarchicalEdgeBundles(phylo, graph, type = "fan", beta = 0.95,
                            args.lines = list(col = alp("steelblue", 90)))
plotHierarchicalEdgeBundles(phylo, graph, type = "fan", beta = 0.85,
                            args.lines = list(col = alp("steelblue", 90)))

# Extra control of plotting and debugging
par(mfrow = c(1,2))
plot(phylo, type = "unrooted")
plotHierarchicalEdgeBundles(phylo, graph, type = "unrooted", beta = 0.8,
                            v.use.only = 1, debug = FALSE,
                            args.lines = list(col = alp("red", 1), lwd = 2))
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\references{
Holten, Danny. "Hierarchical edge bundles: Visualization of adjacency
  relations in hierarchical data." Visualization and Computer Graphics, IEEE
  Transactions on 12.5 (2006): 741-748.
}
\seealso{
See \code{\link[ape]{plot.phylo}}.
}

