% Generated by roxygen2 (4.0.2): do not edit by hand
\name{straightenedBezier}
\alias{straightenedBSpline}
\alias{straightenedBezier}
\title{Straightened Spline curves}
\usage{
straightenedBezier(x, y, beta = 0.5, n.evals = 100)

straightenedBSpline(x, y, order = 4, beta = 0.5, n.evals = 100)
}
\arguments{
\item{x}{A numeric vector giving the x-coordinates of the control points.
Alternatively, it can be a \code{list} (or \code{data.frame}) containing
two vectors of control points.}

\item{y}{A numeric vector giving the y-coordinates of the control points.}

\item{beta}{A numeric values that control amount of straightening.
A \code{beta} of \eqn{1} is a normal spline curve and where as a value of
\eqn{0} is a straight line.}

\item{n.evals}{The number of evaluations of the Bezier curve.}

\item{order}{The order of the B-spline. See \code{\link{bSpline}}.}
}
\value{
Returns a \code{list} of length 2 with entries \code{x} and \code{y} of
  x and y coordinates corresponding to the evaluations of the B-spline.
}
\description{
Function for straightning Bezier or B-spline curves.
}
\note{
The straightening is done by modifying the control points as described
  in the reference.
}
\examples{
# Bezier curves
x <- list(x = cumsum(rnorm(7)),
          y = cumsum(rnorm(7)))
plot(x, type = "b", col = "grey", pch = 16)
for (b in seq(0.1, 1, l = 9))
  lines(Bmisc:::straightenedBezier(x, beta = b))
# B-splines
x <- cumsum(rnorm(7))
y <- cumsum(rnorm(7))
plot(x, y, type = "b", col = "grey", pch = 16)
for (b in seq(0.1, 0.9, l = 9))
  lines(Bmisc:::straightenedBSpline(x, y, beta = b))
}
\author{
Anders Ellern Bilgrau
}
\references{
Holten, Danny. "Hierarchical edge bundles: Visualization of adjacency
  relations in hierarchical data." Visualization and Computer Graphics,
  IEEE Transactions on 12.5 (2006): 741-748.
}
\keyword{internal}

